Support updates for some network attributes
