+++ libvirt/provider_test.go
@@ -1,12 +1,17 @@
 package libvirt
 
 import (
+	"fmt"
 	"os"
+	"reflect"
 	"strings"
 	"testing"
 
+	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/terraform"
+	"github.com/libvirt/libvirt-go"
+	"github.com/libvirt/libvirt-go-xml"
 	"github.com/terraform-providers/terraform-provider-ignition/ignition"
 )
 
@@ -40,4 +45,194 @@ func testAccPreCheck(t *testing.T) {
 func testAccEnabled() bool {
 	v := os.Getenv("TF_ACC")
 	return v == "1" || strings.ToLower(v) == "true"
+}
+
+// //////////////////////////////////////////////////////////////////
+// network
+// //////////////////////////////////////////////////////////////////
+
+// testAccCheckNetworkExists checks that the network exists
+func testAccCheckNetworkExists(name string, network *libvirt.Network) resource.TestCheckFunc {
+	return func(state *terraform.State) error {
+
+		rs, err := getResourceFromTerraformState(name, state)
+		if err != nil {
+			return err
+		}
+
+		virConn := testAccProvider.Meta().(*Client).libvirt
+		fmt.Printf("%s", virConn)
+		networkRetrived, err := virConn.LookupNetworkByUUIDString(rs.Primary.ID)
+		if err != nil {
+			return err
+		}
+
+		realID, err := networkRetrived.GetUUIDString()
+		if err != nil {
+			return err
+		}
+
+		if realID != rs.Primary.ID {
+			return fmt.Errorf("Libvirt network not found")
+		}
+
+		*network = *networkRetrived
+
+		return nil
+	}
+}
+
+// testAccCheckLibvirtNetworkDestroy checks that the network has been destroyed
+func testAccCheckLibvirtNetworkDestroy(s *terraform.State) error {
+	virtConn := testAccProvider.Meta().(*Client).libvirt
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "libvirt_network" {
+			continue
+		}
+		_, err := virtConn.LookupNetworkByUUIDString(rs.Primary.ID)
+		if err == nil {
+			return fmt.Errorf(
+				"Error waiting for network (%s) to be destroyed: %s",
+				rs.Primary.ID, err)
+		}
+	}
+	return nil
+}
+
+// testAccCheckDNSHosts checks the expected DNS hosts in a network
+func testAccCheckDNSHosts(name string, expected []libvirtxml.NetworkDNSHost) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+
+		virConn := testAccProvider.Meta().(*Client).libvirt
+		networkDef, err := getNetworkDef(s, name, *virConn)
+		if err != nil {
+			return err
+		}
+		if networkDef.DNS == nil {
+			return fmt.Errorf("DNS block not found in networkDef")
+		}
+		actual := networkDef.DNS.Host
+		if len(expected) != len(actual) {
+			return fmt.Errorf("len(expected): %d != len(actual): %d", len(expected), len(actual))
+		}
+		for _, e := range expected {
+			found := false
+			for _, a := range actual {
+				if reflect.DeepEqual(a.IP, e.IP) && reflect.DeepEqual(a.Hostnames, e.Hostnames) {
+					found = true
+					break
+				}
+			}
+			if !found {
+				return fmt.Errorf("Unable to find:%v in: %v", e, actual)
+			}
+		}
+		return nil
+	}
+}
+
+// testAccCheckLibvirtNetworkDhcpStatus checks the expected DHCP status
+func testAccCheckLibvirtNetworkDhcpStatus(name string, expectedDhcpStatus string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		virConn := testAccProvider.Meta().(*Client).libvirt
+		networkDef, err := getNetworkDef(s, name, *virConn)
+		if err != nil {
+			return err
+		}
+		if expectedDhcpStatus == "disabled" {
+			for _, ips := range networkDef.IPs {
+				// &libvirtxml.NetworkDHCP{..} should be nil when dhcp is disabled
+				if ips.DHCP != nil {
+					fmt.Printf("%#v", ips.DHCP)
+					return fmt.Errorf("the network should have DHCP disabled")
+				}
+			}
+		}
+		if expectedDhcpStatus == "enabled" {
+			for _, ips := range networkDef.IPs {
+				if ips.DHCP == nil {
+					return fmt.Errorf("the network should have DHCP enabled")
+				}
+			}
+		}
+		return nil
+	}
+}
+
+// testAccCheckLibvirtNetworkBridge checks the bridge exists and has the expected properties
+func testAccCheckLibvirtNetworkBridge(resourceName string, bridgeName string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		virConn := testAccProvider.Meta().(*Client).libvirt
+		networkDef, err := getNetworkDef(s, resourceName, *virConn)
+		if err != nil {
+			return err
+		}
+
+		if networkDef.Bridge == nil {
+			return fmt.Errorf("Bridge type of network should be not nil")
+		}
+
+		if networkDef.Bridge.Name != bridgeName || networkDef.Bridge.STP != "on" {
+			fmt.Printf("%#v", networkDef)
+			return fmt.Errorf("fail: network brigde property were not set correctly")
+		}
+
+		return nil
+	}
+}
+
+// testAccCheckLibvirtNetworkDNSForwarders checks the DNS forwarders in the libvirt network
+func testAccCheckLibvirtNetworkDNSForwarders(name string, expected []libvirtxml.NetworkDNSForwarder) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+
+		virConn := testAccProvider.Meta().(*Client).libvirt
+
+		networkDef, err := getNetworkDef(s, name, *virConn)
+		if err != nil {
+			return err
+		}
+		if networkDef.DNS == nil {
+			return fmt.Errorf("DNS block not found in networkDef")
+		}
+		actual := networkDef.DNS.Forwarders
+		if len(expected) != len(actual) {
+			return fmt.Errorf("len(expected): %d != len(actual): %d", len(expected), len(actual))
+		}
+		for _, e := range expected {
+			found := false
+			for _, a := range actual {
+				if reflect.DeepEqual(a, e) {
+					found = true
+					break
+				}
+			}
+			if !found {
+				return fmt.Errorf("Unable to find %v in %v", e, actual)
+			}
+		}
+		return nil
+	}
+}
+
+// testAccCheckLibvirtNetworkLocalOnly checks the local-only property of the Domain
+func testAccCheckLibvirtNetworkLocalOnly(name string, expectLocalOnly bool) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+
+		virConn := testAccProvider.Meta().(*Client).libvirt
+
+		networkDef, err := getNetworkDef(s, name, *virConn)
+		if err != nil {
+			return err
+		}
+		if expectLocalOnly {
+			if networkDef.Domain == nil || networkDef.Domain.LocalOnly != "yes" {
+				return fmt.Errorf("networkDef.Domain.LocalOnly is not true")
+			}
+		} else {
+			if networkDef.Domain != nil && networkDef.Domain.LocalOnly != "no" {
+				return fmt.Errorf("networkDef.Domain.LocalOnly is true")
+			}
+		}
+		return nil
+	}
 }
